scsi: don't send timeout request to EH anymore
For timeout request, complete it directly without sending it to
LLD driver for processing. We increment the ref_count of request
which will prevent the request from releasing. Though this will
lead to memory leak to some extent, but can avoid system panic if
the LLD driver begin to handle the request again. Meanwhile, we
also don't free scsi_cmnd.

Signed-off-by: Sha Zhengju <handai.szj@taobao.com>