cgroups: new resource counter inheritance API
Provide an API to inherit a counter value from a parent.  This can be
useful to implement cgroup.clone_children on a resource counter.

Still the resources of the children are limited by those of the parent, so
this is only to provide a default setting behaviour when clone_children is
set.

Signed-off-by: Frederic Weisbecker <fweisbec@gmail.com>
Signed-off-by: Wenzhang <wenzhang.zzy@alibaba-inc.com>
Reviewed-by: Zhi Yong Wu <zhiyong.wzy@alibaba-inc.com>