cgroups: pull up res counter charge failure interpretation to caller
res_counter_charge() always returns -ENOMEM when the limit is reached and
the charge thus can't happen.

However it's up to the caller to interpret this failure and return the
appropriate error value.  The nproc counter of kcounter subsystem will need to
report the user that a fork() has been cancelled because of some limit reached,
not because we are too short on memory.

Fix this by returning -1 when res_counter_charge() fails.

Signed-off-by: Frederic Weisbecker <fweisbec@gmail.com>
Signed-off-by: Wenzhang <wenzhang.zzy@alibaba-inc.com>
Reviewed-by: Zhi Yong Wu <zhiyong.wzy@alibaba-inc.com>