cgroups: Add a kcounter subsystem
Add a new subsystem to limit usage of various types of resource
for a cgroup, by enforcing a upper bound upon the corresponding
resource counter. It behaves like the counterpart of user land
ulimit in the scope of a cgroup.

Currently nproc counter is the only consumer of kcounter
framework. For nproc counter, the user can set an upper bound
limit that is checked every time a task forks in a cgroup or is
moved into a cgroup with that subsystem binded.

The primary goal is to protect against forkbombs that explode
inside a container. The traditional NR_PROC rlimit is not
efficient in that case because if we run containers in parallel
under the same user, one of these could starve all the others
by spawning a high number of tasks close to the user wide limit.

This is a prevention against forkbombs, so it's not deemed to
cure the effects of a forkbomb when the system is in a state
where it's not responsive. It's aimed at preventing from ever
reaching that state and stop the spreading of tasks early.
While defining the limit on the allowed number of tasks, it's
up to the user to find the right balance between the resource
its containers may need and what it can afford to provide.

As it's totally dissociated from the rlimit NR_PROC, both
can be complementary: the cgroup nproc counter can set an upper
bound per container and the rlmit can be an upper bound on the
overall set of containers.

Also nproc counter can be used to kill all the tasks in a cgroup
without races against concurrent forks, by setting the limit of
tasks to 0, any further forks can be rejected. This is a good
way to kill a forkbomb in a container, or simply kill any container
without the need to retry an unbound number of times.

Signed-off-by: Frederic Weisbecker <fweisbec@gmail.com>
Signed-off-by: Wenzhang <wenzhang.zzy@alibaba-inc.com>
Reviewed-by: Zhi Yong Wu <zhiyong.wzy@alibaba-inc.com>