tcp: extract code to compute SYNACK RTT
In preparation for adding another spot where we compute the SYNACK
RTT, extract this code so that it can be shared.

[Note from Zheng Liu]
We should backport this patch before 13th patch but we don't do that
because it is only a minor change.  However, we must backport it now
due to the next patch needs it. :(

Signed-off-by: Neal Cardwell <ncardwell@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Zheng Liu <wenqing.lz@taobao.com>