blkcg: still charge bps for sequential bios
In blkcg io throttle, we have changed its behavior for sequential bios:
if the current bio is found to be sequential with the previous one, don't
charge it (both iops and bps) and let it go down directly. This can make
io-scheduler merge them more easily.

But now we want to keep charging bps for these bios, so some changes are needed:
1) move the bio sequential judgement from tg_may_dispatch() to tg_with_in_iops_limit(),
so bps check is still needed
2) only iops is not charged if sequential detected
3) if workqueue is dispatching bio, also check bps limits together with sequential

Signed-off-by: Sha Zhengju <handai.szj@alibaba-inc.com>